common, db: move murmur3 port to common/murmur3, export pair hash
